% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toronto.community.assets-package.R
\docType{package}
\name{toronto.community.assets-package}
\alias{toronto.community.assets}
\alias{toronto.community.assets-package}
\title{toronto.community.assets: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/context-dependent/toronto.community.assets}
  \item Report bugs at \url{https://github.com/context-dependent/toronto.community.assets/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
