% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toronto.community.assets-package.R
\docType{package}
\name{toronto.community.assets-package}
\alias{toronto.community.assets}
\alias{toronto.community.assets-package}
\title{toronto.community.assets: Provides Toronto Community Asset Database in R}
\description{
The Toronto Community Assets Database is a rough list of all 'community assets' in Toronto. Community assets include social services, community programming, cultural institutions, justice programs, and other services, locations, and programs that are available to the public.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/context-dependent/toronto.community.assets}
  \item Report bugs at \url{https://github.com/context-dependent/toronto.community.assets/issues}
}

}
\author{
\strong{Maintainer}: Thomas McManus \email{tmcmanus@blueprint-ade.ca}

}
\keyword{internal}
